Refactor logic of data lists' length comparison
The logic of data list length comparison in task.py has been refactored. Previously it was checking the length of reactions list against message list, but now it checks the length of hour list against message list. The purpose of this refactor is to better handle data discrepancies and potentially prevent a data mismatch in the future.