Refactor functions and handling reactions data
Updated the way the driver accesses the URL and refactored how the reaction data is handled. The changes will ensure that reactions align correctly with their corresponding messages, even if there are fewer reactions than messages. A more reliable method was also added to accommodate cases where the reactions' data might be empty or not equal to the length of messages, preventing the program from potential out-of-index errors.