Refactor task module to streamline channel handling
The revision simplifies the task execution by removing the need to pass the list of channels to the main_task function. Now, the script acquires the list of channels directly from the browser, using the newly imported task_utils module. This streamlines the process and increases the adaptability of the code.