Implement dynamic short time calculation in constants
Updated constants.py to include dynamic calculation of 'SHORT_TIME' instead of a fixed value. Integration of the 'botasaurus' library was carried out to define the 'MPL' constant, which is later used to adjust 'SHORT_TIME' based on parallel browsers. This will allow a more responsive and adaptive timeout configuration.